Changed wooden wheel into simply Wheel™, now it's made of iron and is a bit bigger Fixed rendering issues for belts Re-balanced all rotary power machines Fixed skycrate and skycart stations, added aabb and improved minecart handling

+++ src/main/java/pl/pabilo8/immersiveintelligence/api/rotary/IMotorBeltConnector.java
@@ -45,5 +45,5 @@ public interface IMotorBeltConnector
 	EnumFacing.Axis getConnectionAxis();
 
 	//Radius of the conveyor in 1/16 of a block
-	int getRadius();
+	float getRadius();
 }
